+++ scripts/lib/svc_md_to_html.py
@@ -8,7 +8,7 @@ from pathlib import Path
 
 from pypandoc import convert_text
 
-from scripts.lib.utils.utils import get_md_text
+from scripts.lib.utils.common_utils import get_md_text
 
 
 def convert_md_to_html(md_text: str, output_path: Path, is_strip_wrapper: bool = False) -> None:
