+++ scripts/lib/svc_md_to_html_text.py
@@ -20,7 +20,7 @@ def convert_md_to_html_text(md_text: str, is_strip_wrapper: bool = False) -> str
         Exception: If conversion fails
     """
     # Process Markdown text
-    from scripts.lib.utils.utils import get_md_text
+    from scripts.lib.utils.common_utils import get_md_text
     processed_md = get_md_text(md_text, is_strip_wrapper=is_strip_wrapper)
     
     # Convert to HTML
