+++ scripts/lib/svc_md_to_latex.py
@@ -20,7 +20,7 @@ def convert_md_to_latex(md_text: str, output_path: Path, is_strip_wrapper: bool 
         Exception: If conversion fails
     """
     # Process Markdown text
-    from scripts.lib.utils.utils import get_md_text, parse_md_to_tables
+    from scripts.lib.utils.common_utils import get_md_text, parse_md_to_tables
     processed_md = get_md_text(md_text, is_strip_wrapper=is_strip_wrapper)
     
     # Parse Markdown tables
