+++ scripts/lib/svc_md_to_linked_image.py
@@ -13,7 +13,7 @@ import httpx
 import markdown
 from bs4 import BeautifulSoup
 
-from .utils.utils import get_md_text
+from .utils.common_utils import get_md_text
 
 # MIME type mapping
 MIME_TYPE_MAP = {
