+++ scripts/lib/svc_md_to_md.py
@@ -20,7 +20,7 @@ def convert_md_to_md(md_text: str, output_path: Path, is_strip_wrapper: bool = F
         Exception: If conversion fails
     """
     # Process Markdown text
-    from scripts.lib.utils.utils import get_md_text
+    from scripts.lib.utils.common_utils import get_md_text
     processed_md = get_md_text(md_text, is_strip_wrapper=is_strip_wrapper)
     
     # Write to output file
