+++ scripts/lib/svc_md_to_pdf.py
@@ -8,7 +8,7 @@ from pathlib import Path
 
 from xhtml2pdf import pisa
 
-from scripts.lib.utils.utils import (
+from scripts.lib.utils.common_utils import (
     contains_chinese,
     contains_japanese,
     convert_markdown_to_html,
