+++ scripts/lib/svc_md_to_xml.py
@@ -23,7 +23,7 @@ def convert_md_to_xml(md_text: str, output_path: Path, is_strip_wrapper: bool = 
         Exception: If conversion fails
     """
     # Process Markdown text
-    from scripts.lib.utils.utils import get_md_text
+    from scripts.lib.utils.common_utils import get_md_text
     processed_md = get_md_text(md_text, is_strip_wrapper=is_strip_wrapper)
     
     # Convert to XML
