+++ scripts/lib/utils/common_utils.py
